fix: move getStreamTypeConfig to index
